+++ internal/tftime/provider.go
@@ -7,6 +7,7 @@ import (
 func Provider() *schema.Provider {
 	return &schema.Provider{
 		ResourcesMap: map[string]*schema.Resource{
+			"time_offset":   resourceTimeOffset(),
 			"time_rotating": resourceTimeRotating(),
 			"time_static":   resourceTimeStatic(),
 		},